requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-telegram-bot>=20.3
fpdf2>=2.7.4
openpyxl>=3.1.0  # Required for pandas Excel support
xlrd >= 2.0.1
# Optional dependencies